Get rid of "xml" and "lib" keywords
We now support arbitrary SVG (but it still has to fit on one line). this
will allow us to use markers and clips. The drawback is that we no
longer know which of these are actually required for a particular map.
For the moment, we include all of these terms in our definitions. We
might want to create a dependency graph. We would have to follow all the
xlink:href attributes in use elements, all the marker-start and
marker-end attributes in CSS properties for path, line, polyline and
polygon elements, and more. This is tedious and therefore this is not
done.