Looks like there is a way to split types
This change does not change how data is parsed, because at parse time,
we don't know whether "green" is going to be a background color or
not.

When rendering, however, we know because "green" ended up defining
(hex) attributes. So now when rendering a hex, we can look at all its
types and put the attributes into the background and keep the
non-attributes in the foreground (above the lines, i.e. above the
rivers and roads).