Editable maps
This commit allows you to edit the UWP, generate a map, and link back
to the UWP with no id. This means that in those cases where the id is
lost, the map data itself is appended to the query. Sadly, this
generates a lot of problems.

The first problem is that an URL containing the complete map data is
hardly "shareable". This is the URL for a single system:
http://localhost:3000/uwp?map=Inedgeus++++++++++0101+D7A5579-8++++++++G++Fl+NI++++++++++A%0D%0A

In addition to that, URLs are limited in size, so where as it works
for subsectors, it won't work for sectors. They are simply too long. I
would have to store the edited maps on the server and I like the
current stateless implementation.

One additional issue has been the use of & to separate parameters in
links for the link in the SVG itself. If you switch to CT (also
something I'd have to support), then the source URL for the UWP is
something like http://localhost:3000/uwp?classic=1&map=... which is
invalid SVG because of the &, so I'd have to look at the escaping
rules.