Rewrite of some Traveller code
I needed some functions like nearby and distance elsewhere so I moved
it into a separate package and decided to rewrite the classes to use
Moose instead of Class::Struct so that I could have these utility
functions in a role.

After generating a system, systems are now renamed based on a
"culture" that gets started all over the map. The result is that
multiple "languages" are used for the system names and the backgrounds
are tinted to indicate where these historic boundaries are.

These "culture" markers also led to a non-standard extension of the
UWP, unfortunately. The "culture" is now a number in brackets at the
end of the line, which necessitated some changes to the parser.