hex-describe: fixes to nearby/other rules
Previously, *[nearby foo]* resulted in the brackets and the link with
the coordinates to also be inside the em element. No more!

Treasure hunters added to the big table in contrib.
Necromancers now have the Agrimach spellbock.

Various patrons are now defined, including alchemist patrons. All the
things of interest to alchemists are now also defined as patron
quests. Treasure hunters now want to go on these quests.