text-mapper: improve corridors
Start to search more (otherwise we would rarely create a corridor that
only touched diagonally with a 2x2 room) space for a corridor. This
led to more overhang (overshooting at both ends). So now add code that
checks whether we've started too early or have continued for too long
and trim the excess corridor.