hex-describe: fix load order for the table
When using get_table, the load order is no changed. Frist the default
table is loaded, and then the text area is appended. In parse_table,
this now triggers the "reset" code if a table name is reused. These
table are no longer merged. Instead, the table from the text area
takes precedence and overrides the default.