hex-describe: change how URLs are initialised
The previous mechanism didn't work when restarting the process via
Monit and hypnotoad: it kept using development mode. The code was
changed to no longer rely on the mode and simply override the default
URLs with values in the conf file. The content for that conf file
which makes the local-test target in the Makefile work is this:

{
  hex_describe_url => 'http://localhost:3000',
  text_mapper_url => 'http://localhost:3010',
  face_generator_url => 'http://localhost:3020',
}