Handle &#45; in map input & document license
We’re translating - to &#45; for the commented source code at the end
where two dashes would be illegal. If somebody then copied the source
code and pasted it back into the Text Mapper, we need to handle this.

Document that the license tag has fixed coordinates supplied by the
user (or the library) and thus if the map does not include the hex
0101 it might not be visible.