schroeder: reworking the dungeons
Moving some of the HTML used to generate the room lists into their own
little tables for easier testing.

Removed the italics for the room descriptions.

Rewrote the traps to be more precise: provide better warnings for
players, add attack rolls, specified damage, qualified poison effects.
Removed the bold maroon for traps and used simple bold just as for
monsters, reasoning that bold = dangers the referee has to know about.

Rewrote the hiding places of unguarded treasure. Firstly, putting it
ahead of the treasure description instead of in brackets after the
treasure, by rewriting it into a sequence of discoveries, and by
proding a more specific list of hiding places.