traveller: population-0 and captains' associations
Wrote special case tables for population-0 and various forms of
government (government-0 to government-5). For one of the entries, I
wrote a "captains' association" to which some captains may belong.

Added a crew to scout bases. Added an association of former scouts at
this base and made that into a "captains' association" as well.