hex-describe: foo-n defines key "foo" to be "n"
This is useful for Traveller, where we can now have "law-5" defining
the local key "law" to have the value "5", which in turn allows us to
build on it.

Fixed handling of TOP and END such that they don't get printed if
empty.

Fixed the examples ("hills" → "hill", 0102 → 0201) so that they look
better.

Added documentation of the above, as well of the general defining of
locals with the value "1" if the word is present.