text-mapper: set base URL scheme to HTTPS
The problem is that in some cases we want an absolute URL and not a
relative one, e.g. when generating the alpine documentation link. This
is what we don't want when generating the link on the backend:

http://localhost:4009/text-mapper/alpine/document?seed=823579188

The first step is to setup up the reverse proxy such that it passes
along the the name:

ProxyPreserveHost On

This fixes the hostname, but not the scheme:

http://campaignwiki.org/text-mapper/alpine/document?seed=823579188

In theory, the fix ought to be something like this:

RequestHeader set X-Forwarded-Proto "https"

Sadly, it seems to have no effect. I finally decided to go with the
solution proposed in
https://docs.mojolicious.org/Mojolicious/Guides/Cookbook#Rewriting
in order to change scheme if "X-Forwarded-Proto" header is
set (presumably to HTTPS).

app->hook(before_dispatch => sub {
  my $c = shift;
  $c->req->url->base->scheme('https')
      if $c->req->headers->header('X-Forwarded-Proto') } );